fix: delete Stones class & add stones property to Board class
