--- src/main/kotlin/omok/model/Stones.kt
@@ -1,13 +1,0 @@
-package omok.model
-
-class Stones {
-    private val _stones = mutableListOf<Stone>()
-    val stones: List<Stone>
-        get() = _stones.toList()
-
-    fun add(stone: Stone) {
-        _stones.add(stone)
-    }
-
-    fun lastStone(): Stone? = stones.lastOrNull()
-}
